Update to DotNet Core 3.1.3

+++ Master40.SimulationImmutables/Message/FRequestProposalForSetups.fs
@@ -4,5 +4,5 @@ open IJobs
 
 type public FRequestProposalForSetup = {
     Job : IJob
-    SetupId : int64
+    SetupId : int32
 }
